requirements.txt
fastapi
uvicorn
pandas
xgboost
streamlit
pillow
requests
pydantic
scikit-learn==1.6.1
joblib
